Finished the merge, introduced the compression write and the non compression write of the inverted index, changed the length from int to long.

+++ inverted-index/src/main/java/it/unipi/mircv/utils/Utils.java
@@ -27,7 +27,7 @@ public class Utils {
      * @param number Number of which compute the splits.
      * @return Returns the number of splits.
      */
-    public static int splitsLog128(int number){
+    public static int splitsLog128(long number){
         return (int)(Math.floor(Math.log(number) / Math.log(128)) + 1);
     }
 }
